backend/requirements.txt
# ShapeX Backend Dependencies

# Web Framework
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.0
python-multipart==0.0.6
websockets==13.1  # For Studio WebSocket streaming

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# AI & Analysis
anthropic>=0.40.0  # Updated for messages API
openai==1.6.0  # Alternative AI provider

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
selenium==4.16.0
playwright==1.40.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Scheduling
APScheduler==3.10.4

# Telegram Integration
python-telegram-bot==20.7

# Payment Processing
stripe==8.0.0

# Environment & Config
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Utilities
pytrends==4.9.2  # Google Trends
pydantic-settings==2.1.0
httpx~=0.25.2  # Compatible with python-telegram-bot
aiohttp==3.9.1
